+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Add sanitization for nonProxyHosts RegEx patterns
     - Fix bug with malformed file during S3 upload
     - Added periodic closure of sockets closed by the remote end (snowflakedb/snowflake-jdbc#2481).
+    - Add internal API usage telemetry tracker
 
 - v4.0.1
     - Add /etc/os-release data to Minicore telemetry